+++ lib/save_us_starmax_platform_interface.dart
@@ -38,7 +38,19 @@ abstract class SaveUsStarmaxPlatform extends PlatformInterface {
     throw UnimplementedError();
   }
 
+  Future<Uint8List?> reset() {
+    throw UnimplementedError();
+  }
+
+  Future<Uint8List?> setState(Map<String, dynamic> param) {
+    throw UnimplementedError();
+  }
+
   Future<Uint8List?> setTime() {
+    throw UnimplementedError();
+  }
+
+  Future<Uint8List?> setUserInfo(Map<String, dynamic> param) {
     throw UnimplementedError();
   }
 
@@ -46,47 +58,55 @@ abstract class SaveUsStarmaxPlatform extends PlatformInterface {
     throw UnimplementedError();
   }
 
+  Future<Uint8List?> setNotDisturb(Map<String, dynamic> param) {
+    throw UnimplementedError();
+  }
+
+  Future<Uint8List?> getNotDisturb() {
+    throw UnimplementedError();
+  }
+
   Future<Uint8List?> getSportHistory() {
     throw UnimplementedError();
   }
 
-  Future<Uint8List?> getValidHistoryDates() {
+  Future<Uint8List?> getValidHistoryDates(Map<String, dynamic> param) {
     throw UnimplementedError();
   }
 
-  Future<Uint8List?> getStepHistory() {
+  Future<Uint8List?> getStepHistory(Map<String, dynamic> param) {
     throw UnimplementedError();
   }
 
-  Future<Uint8List?> getHeartRateHistory() {
+  Future<Uint8List?> getHeartRateHistory(Map<String, dynamic> param) {
     throw UnimplementedError();
   }
 
-  Future<Uint8List?> getBloodPressureHistory() {
+  Future<Uint8List?> getBloodPressureHistory(Map<String, dynamic> param) {
     throw UnimplementedError();
   }
 
-  Future<Uint8List?> getBloodOxygenHistory() {
+  Future<Uint8List?> getBloodOxygenHistory(Map<String, dynamic> param) {
     throw UnimplementedError();
   }
 
-  Future<Uint8List?> getPressureHistory() {
+  Future<Uint8List?> getPressureHistory(Map<String, dynamic> param) {
     throw UnimplementedError();
   }
 
-  Future<Uint8List?> getMetHistory() {
+  Future<Uint8List?> getMetHistory(Map<String, dynamic> param) {
     throw UnimplementedError();
   }
 
-  Future<Uint8List?> getTempHistory() {
+  Future<Uint8List?> getTempHistory(Map<String, dynamic> param) {
     throw UnimplementedError();
   }
 
-  Future<Uint8List?> getMaiHistory() {
+  Future<Uint8List?> getMaiHistory(Map<String, dynamic> param) {
     throw UnimplementedError();
   }
 
-  Future<Uint8List?> getBloodSugarHistory() {
+  Future<Uint8List?> getBloodSugarHistory(Map<String, dynamic> param) {
     throw UnimplementedError();
   }