+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-## 0.0.8
-## 0.0.9
+## 1.0.4
+Implemented to retrieve history data
+## 1.0.3
+Add example
+## 1.0.2
+## 1.0.1
+Add Reset function
 ## 1.0.0
+## 0.0.9
+## 0.0.8